Task 9: split bootstrap.js into focused modules
Extract three responsibilities from the 376-line bootstrap.js:
- src/configLoader.js: fetch heroName/monsterList/dungeons JSON at startup
- src/storeSetup.js: create Redux store and bind all 8 services
- src/api.js: build the game API object consumed by React via GameContext

bootstrap.js is now a ~40-line orchestrator that sequences these steps.
